FINAL! Please read description.
This is the final commit from me. I've checked everything and added a
comment so it's ready to download and play as it's using static dummy
data. The SQL export is there if you want to setup and use your own but
you'll have to config it through the PHP files. Ben